download-sbom: always create SBOMS_DIR
Currently, when there are no SBOMs to download->upload, the pipeline
fails because the upload-sbom script runs on a non-existing directory.

Even if there are no SBOMs to download, create an empty SBOMS_DIR. This
should serve as an indication for the upload-sbom script that everything
ran correctly, but there really are no SBOMs to upload.

Signed-off-by: Adam Cmiel <acmiel@redhat.com>