comment out ACS env vars in GitHub Actions
The automatically generated GitHub Actions workflows include a step
which checks that exposed environment variables are actually set.

`REKOR_HOST` and `TUF_MIRROR` should not be required and there is
currently no way to easily mark the env vars as not required for the
required check step.
Comment them out and let users uncomment them when they want to use an
external instance